+++ include-xpl/xpl_texture_atlas.h
@@ -34,10 +34,10 @@ typedef struct xpl_texture_atlas {
 	xpl_texture_atlas_node_t    *nodes;     // utlist
 	int                         node_count;
 
-	int							width;
-	int							height;
+	size_t						width;
+	size_t						height;
 #ifndef XPL_PLATFORM_IOS
-	int							depth;      // buffer depth in bytes (4 = 32 bit)
+	size_t						depth;      // buffer depth in bytes (4 = 32 bit)
 #endif
 
 	int							used;       // surface size used
@@ -48,7 +48,7 @@ typedef struct xpl_texture_atlas {
 
 } xpl_texture_atlas_t;
 
-xpl_texture_atlas_t *xpl_texture_atlas_new(const int width, const int height, const int depth);
+xpl_texture_atlas_t *xpl_texture_atlas_new(const size_t width, const size_t height, const size_t depth);
 void xpl_texture_atlas_destroy(xpl_texture_atlas_t **ppatlas);
 
 // Assemble texture atlas from multiple bitmaps.
